ai_research_agent/requirements.txt
openai
python-dotenv
matplotlib
letta-client
